split Parsing.swift into multiple files, enable JSOONEncdedBody with encodable, remove header from files
