+++ Sources/Body/FormEncodedBody.swift
@@ -1,11 +1,3 @@
-//
-//  FormEncodedBody.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 /// A type representing a form encoded HTTP request body.
