+++ Sources/Call.swift
@@ -1,11 +1,3 @@
-//
-//  Call.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 /// A type representing a call to a Web API endpoint.
