+++ Sources/Convenience/AnyClient.swift
@@ -1,11 +1,3 @@
-//
-//  AnyClient.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 /// The default implementation of `Client`.
