+++ Sources/Convenience/BasicAuthorization.swift
@@ -1,11 +1,3 @@
-//
-//  BasicAuthorization.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 public struct BasicAuthorization {
