+++ Sources/Convenience/Result+Extensions.swift
@@ -1,11 +1,3 @@
-//
-//  Result+Extensions.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 extension Result {
