+++ Sources/Error/StatusCodeError.swift
@@ -1,11 +1,3 @@
-//
-//  StatusCodeError.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 /// Wraps an HTTP error code.
