+++ Sources/HTTPMethod.swift
@@ -1,11 +1,3 @@
-//
-//  HTTPMethod.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 /// An enum containing all HTTPMethods defined in RFC 2616
