+++ Sources/ResponseValidator/ResponseValidator.swift
@@ -1,11 +1,3 @@
-//
-//  ResponseValidator.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 /// A type responsible for validating the result produced by a
