+++ Sources/Testing/FakeResultProvider.swift
@@ -1,11 +1,3 @@
-//
-//  FakeResultProvider.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 public protocol FakeResultProvider {
