+++ Sources/Testing/FakeSession.swift
@@ -1,11 +1,3 @@
-//
-//  FakeSession.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 public class FakeSession<C: Client>: Session<C> {
