+++ Sources/URLRequestEncodable.swift
@@ -1,11 +1,3 @@
-//
-//  URLRequestEncodable.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 /// A type that can transform itself into an `URLRequest`.
