+++ Sources/URLSessionTaskResult.swift
@@ -1,11 +1,3 @@
-//
-//  URLSessionTaskResult.swift
-//  Endpoints
-//
-//  Created by Robin Mayerhofer on 27.08.19.
-//  Copyright © 2019 Tailored Apps. All rights reserved.
-//
-
 import Foundation
 
 /// Encapsulates the result produced by a `URLSession`s
